test(gui): refactor E2E handler file names
