+++ mocks/handlers/configs/cppConfig.ts
