+++ mocks/handlers/configs/erroneousConfig.ts
