+++ mocks/handlers/index.ts
@@ -2,6 +2,6 @@
  * @file Exports all handlers within this directory. It expects each file should export a `handlers` array.
  */
 
-export * as cppAssignment from "./configs/cppAssignment";
-export * as erroneousAssignment from "./configs/erroneousAssignment";
+export * as cppConfig from "./configs/cppConfig";
+export * as erroneousConfig from "./configs/erroneousConfig";
 export * as global from "./global";
